hilt app module impl

+++ networkModule/src/main/java/pl/example/networkmodule/apis/ResultApi.kt
@@ -85,7 +85,7 @@ class ResultApi(private val ktorClient: KtorClient) : ResultApiInterface {
         }
     }
 
-    override suspend fun getHb1AcResultById(id: String): Double? {
+    override suspend fun getHb1AcResultById(id: String): Float? {
         val response = client.get("$adress/$resultsEndpoint/hb1ac/$id")
         return if (response.status == HttpStatusCode.OK) {
             if (response.contentType()?.match(ContentType.Application.Json) == true) {
